FAB (+) brings the user to the AddNewFeed activity and save redirects back to the main page, also fixed issues with displaying changes upon refreshes
