+++ MyApplication/app/src/main/AndroidManifest.xml
@@ -20,10 +20,14 @@
             </intent-filter>
         </activity>
 
+        <activity android:name=".view.AddNewFeed" >
+        </activity>
+
         <service
             android:name=".service.NewFeedNotifierService"
             android:enabled="true"
-            android:exported="true"></service>
+            android:exported="true">
+        </service>
     </application>
 
 </manifest>
