+++ MyApplication/app/src/main/res/layout/activity_feed_list.xml
@@ -33,6 +33,7 @@
             android:backgroundTint="@color/colorPrimary"
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"
+            android:onClick="addNewFeed"
             />
 
     </android.support.constraint.ConstraintLayout>
